Fix #57: Create YAML configuration files for OpenFAM - Updated based on review comments from Mashood  (Moved test file to test/unit-test/config folder and created sample config file) - Updated CMakeLists.txt to copy the config file during make install.
