--- config/fam_cis_config
@@ -1,20 +1,0 @@
-#CIS Config file
-
-# List of memory servers that is associated with metadata manager.
-memsrv_ip :
-- 0:<ip_address_1>
-- 1:<ip_address_2>
-- 3:<ip_address_3>
-- 2:<ip_address_4>
-
-# metadata manager ip address < this can also be an array right?>
-metadata_ips:
-- 0:<md_ ip_address_1>
-- 1:<md_ ip_address_2>
-           
-# Port to be used to communicate with memory server.
-grpc_port_to_memsrv : xxxx
-
-# Port to be used to communicate with metadata server.
-grpc_port_to_metadatasrv : xxxx
-
