--- config/fam_metadata_config
@@ -1,14 +1,0 @@
-#CIS Config file
-
-# Metadata manager: default is radixtree. Another option is etcd.
-metadata_manager : "radixtree"
- 
-# List of memory servers that is associated with metadata manager.
-memsrv_ip :
-         - 0:ip_address_1
-         - 1:ip_address_2
-         - 3:ip_address_3
-         - 2:ip_address_4
-# Path where metadata info is stored. By default, it will be placed in /dev/shm. 
-metadata_path: /dev/shm/`whoami`
-
