+++ test/unit-test/CMakeLists.txt
@@ -28,3 +28,4 @@ add_subdirectory(allocator)
 add_subdirectory(fam-api)
 add_subdirectory(metadata)
 add_subdirectory(bitmap-manager)
+add_subdirectory(config)
