+++ test/unit-test/config/CMakeLists.txt
@@ -1,6 +1,6 @@
  #
  # CMakeLists.txt
- # Copyright (c) 2019-20 Hewlett Packard Enterprise Development, LP. All rights reserved.
+ # Copyright (c) 2020 Hewlett Packard Enterprise Development, LP. All rights reserved.
  # Redistribution and use in source and binary forms, with or without modification, are permitted provided
  # that the following conditions are met:
  # 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
@@ -22,7 +22,7 @@
 function (add_fam_test file_name)
     add_executable(${file_name} ${file_name}.cpp)
     target_link_libraries(${file_name} openfam gtest yaml-cpp)
-    add_test(NAME ${file_name} COMMAND ${TEST_RUNTIME_BIN} ${TEST_RUNTIME_OPTS} ${CMAKE_CURRENT_BINARY_DIR}/${file_name})
+    add_test(NAME ${file_name} COMMAND ${TEST_RUNTIME_BIN} ${TEST_RUNTIME_OPTS} ${CMAKE_CURRENT_BINARY_DIR}/${file_name} ${CMAKE_CURRENT_BINARY_DIR}/sample_config)
 endfunction()
 
 #add tests
