+++ test/unit-test/config/sample_config
@@ -1,0 +1,19 @@
+# PE config file
+
+# IP address of CIS server 
+cis_ip: 127.0.0.1
+
+# Port to be used by OpenFam grpc memory allocator 
+grpc_port : 8787
+
+# Libfabric Port to be used by OpenFam memory allocator operation
+libfabric_port : 7500
+
+# Libfabric provider to be used by OpenFam libfabric datapath operations; default is sockets. It could be verbs/omnipath
+provider : "sockets"
+
+# Allocator to be used, default is Grpc, Other option is NVMM.
+allocator : "grpc"
+
+# Fam thread model: default is FAM_THREAD_SERIALIZE. Other option is FAM_THREAD_MULTIPLE.
+thread_model: FAM_THREAD_SERIALIZE
